Fully consolidate the two base node displays
