+++ ee/vellum_cli/tests/test_push.py
@@ -194,7 +194,7 @@ def test_push__dry_run_option_returns_report(mock_module, vellum_client):
 from typing import Dict
 from vellum.workflows import BaseWorkflow
 from vellum.workflows.nodes import BaseNode
-from vellum_ee.workflows.display.nodes import BaseNodeDisplay
+from vellum_ee.workflows.display.nodes.base_node_display import BaseNodeDisplay
 
 class NotSupportedNode(BaseNode):
     pass
