+++ ee/vellum_ee/workflows/display/nodes/vellum/try_node.py
@@ -13,7 +13,6 @@ from vellum_ee.workflows.display.nodes.base_node_vellum_display import BaseNodeV
 from vellum_ee.workflows.display.nodes.get_node_display_class import get_node_display_class
 from vellum_ee.workflows.display.nodes.types import NodeOutputDisplay
 from vellum_ee.workflows.display.nodes.utils import raise_if_descriptor
-from vellum_ee.workflows.display.nodes.vellum.base_node import BaseNodeDisplay as GenericBaseNodeDisplay
 from vellum_ee.workflows.display.types import WorkflowDisplayContext
 
 _TryNodeType = TypeVar("_TryNodeType", bound=TryNode)
@@ -36,14 +35,14 @@ class BaseTryNodeDisplay(BaseNodeVellumDisplay[_TryNodeType], Generic[_TryNodeTy
             inner_node = subworkflow.graph
         elif inner_node.__bases__[0] is BaseNode:
             # If the wrapped node is a generic node, we let generic node do adornment handling
-            class TryBaseNodeDisplay(GenericBaseNodeDisplay[node]):  # type: ignore[valid-type]
+            class TryBaseNodeDisplay(BaseNodeDisplay[node]):  # type: ignore[valid-type]
                 pass
 
             return TryBaseNodeDisplay().serialize(display_context)
 
         # We need the node display class of the underlying node because
         # it contains the logic for serializing the node and potential display overrides
-        node_display_class = get_node_display_class(BaseNodeVellumDisplay, inner_node)
+        node_display_class = get_node_display_class(BaseNodeDisplay, inner_node)
         node_display = node_display_class()
 
         serialized_node = node_display.serialize(
@@ -70,7 +69,7 @@ class BaseTryNodeDisplay(BaseNodeVellumDisplay[_TryNodeType], Generic[_TryNodeTy
         if not inner_node:
             return super().get_node_output_display(output)
 
-        node_display_class = get_node_display_class(BaseNodeVellumDisplay, inner_node)
+        node_display_class = get_node_display_class(BaseNodeDisplay, inner_node)
         node_display = node_display_class()
         if output.name == "error":
             return inner_node, NodeOutputDisplay(
